requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
pycocotools
scipy
scikit-learn